fix: tests
Modified tests to more effectively test logic. Turns out couldn't just
look for the variable not to be defined at the top of the shell-specific
script but needed to look for it being set immediately after having sourced
the dotfile, which can happen in different places for each of the [4] supported
shells.