fix(push): handle deleted environments on floxhub (#1611)
`flox push` will fetch the latest changes from floxhub prior to pushing
to detect diverged branches.
If the environment was removed upstream, the fetch failed. As a result,
environments that were deleted before can not be pushed again. With this
PR flox will ignore when git is unable to fetch the upstream branch due
to it missing and will print a warning instead.

Since this routine is a an internal library detail, its in turn an
internal non-user focused warning. Alternatively the warning could be
silenced further or implemented as a "precheck" that is called from the
frontend crate.