fix(search): Search suggestions for catalog (#1628)
## Proposed Changes

Previously the suggestions were coming from `pkgdb` even when the
catalog was enabled. They now use the same path as the original search.
New integration tests have been added for both code paths.

We agreed in refinement that it was sufficient to plumb the catalog
logic through to `DidYouMean` and refactor later when we removed
`pkgdb`. We could group the catalog and pkgdb related arguments in
separate types but it doesn't seem worth it if we've removing this code
soon.

I considered moving the spinners up one or two levels but I think it
makes sense to keep them close to the slow function based on the
problems I ran into before with competing spinners at different levels
of the codebase.

## Release Notes

N/A because it's fixing a bug in something that's not officially
released yet.

---------

Co-authored-by: Zach Mitchell <zmitchell@fastmail.com>
Co-authored-by: Matthew Kenigsberg <matthew@floxdev.com>