fix: Always handle errors encountered when opening environments
Previously, we didn't have a top-level handler for
EnvironmentSelectError, so it fell back to the generic error formatter,
ignoring any custom formatting for nested EnvironmentErrors. We now
explicitly check for the EnvironmentSelectError::Environment variant,
which wraps an EnvironmentError, and we format the nested
EnvironmentError explicitly.

This means that any error that happens while detecting and opening
environments will be handled, including our attempts to synchronize
managed and remote environments.