chore(catalog): Set request user-agent (#1627)
## Proposed Changes

So that we get version information from the catalog service on release:

- https://github.com/flox/catalog-server/issues/113

To set the user-agent we need to construct our own client. The default
timeouts have been cribbed from `catalog::new()`.

Technically it would be better to assert that the header value matches
`^flox-cli/\d\.\d` but we can only do that with a custom `matches()` and
then we don't get useful output when the test fails, like this:

---- providers::catalog::tests::user_agent_set_on_all_requests stdout
----
thread 'providers::catalog::tests::user_agent_set_on_all_requests'
panicked at
/Users/dcarley/.cargo/registry/src/index.crates.io-6f17d22bba15001f/httpmock-0.7.0/src/api/mock.rs:543:35:
assertion `left == right` failed: 0 of 1 expected requests matched the
mock specification, .
Here is a comparison with the most similar non-matching request (request
number 1):

1 : Expected header with name 'user-agent' and value 'fox-cli/1.0.6' to
be present in the request but it wasn't.

------------------------------------------------------------------------------------------
    Expected:					[key=equals, value=equals]		user-agent=fox-cli/1.0.6
Actual (closest match): user-agent=flox-cli/1.0.6

      left: "user-agent=fox-cli/1.0.6"
     right: "user-agent=flox-cli/1.0.6"

**Disable httpmock cookie features**

We're not using these and it cuts down the transient dependencies by a
fair amount, even if we still have to pull in `async-std` alongside
`tokio`.

## Release Notes

N/A