chore(ci): Remove dupe integ-tests from job
The integration tests are being run twice:

- cli-dev: runs them "locally" on two different Runner systems
- nix-build-bats-tests: runs them "remotely" on four different builder systems

I noticed this while reviewing:

- https://github.com/flox/flox/actions/runs/9560120725?pr=1625

`cli-dev` is typically our longest running job in CI and running the
integration tests there only gives us a subset of the coverage that we
get from `nix-build-bats-tests`, so we can save time on the overall CI
run and make failures easier to reason about by only running them in
`nix-build-bats-tests`.

I've plumbed `nix-build-bats-tests` through to `report-failure` at the
same time so that we get the same amount of reporting on `main`. They
should be reliable enough now because we depend on them for PR merges.