fix: show a helpful error when a managed environment has been deleted (#1638)
Previously, we showed a generic failure message when the upstream 
environment was deleted. We now handle that case and show a message 
indicating what's happened and instructing the user how to restore the 
environment.

This also changes the way we handle EnvironmentSelectErrors. We now 
include a check for EnvironmentSelectError in main() alongside the other
error
handler, and handle the EnvironmentSelectError::Environment variant the
same as
ordinary EnvironmentErrors.