feat(catalog): migrate v0 manifests and lockfiles (#1625)
Migrate v0 manifests and lockfiles to v1 in `install`, `uninstall`,
`edit`, `pull`, and `upgrade`.

Note this does not handle migrating manifests that get locked outside of
transactions, as in `activate`, `containerize`, `list`, and `push`.
Those cases will be handled in https://github.com/flox/flox/issues/1624.

Behavior for each command is as follows.

## `install` and `uninstall`

If just a v0 manifest is detected and can be successfully migrated,
migration is performed automatically:
```
⚠️  Detected an old environment manifest.
⬆️  Migrated environment 'name' to version 1.
```
If the environment has a lockfile, the user is prompted:
```
> To edit environment 'name', you must upgrade to environment version 1.
Do you want to upgrade now? Yes
⬆️  Migrated environment 'name' to version 1.
```

If migration of the environment fails, the user is directed to `flox
edit`:
```
⚠️  Detected an old environment manifest.
❌ ERROR: Could not automatically migrate manifest to version 1:

unknown field `prefer-pre-releases`, expected `allow-pre-releases`

Use 'flox edit' to resolve errors and then try again.
```

## `edit`

`edit` behaves the same as `install` and `uninstall`, except if
migration fails, it drops the user into an editor with the old manifest.
If the old manifest is written, the error printed directs the user to
set `version = 1`:
```
❌ ERROR: Modifying version 0 manifests is no longer supported.
Set 'version = 1' in the manifest.
! Continue editing? (Y/n)
```

If the user sets `version = 1` but has another error, that error is
printed:
```
❌ ERROR: Failed to parse manifest:

unknown field `prefer-pre-releases`, expected `allow-pre-releases`
in `options.semver`

! Continue editing? (Y/n)
```

## `upgrade`

`upgrade` always attempts to perform migration. If the environment does
not have a lockfile, `upgrade` prints:
```
⚠️  Detected an old environment version. Attempting to migrate to version 1.
⬆️  Migrated environment 'name' to version 1.
ℹ️  No packages need to be upgraded in environment 'name'.
```

If the environment does have a lockfile, `upgrade` prints:
```
⚠️  Detected an old environment version. Attempting to migrate to version 1 and upgrade packages.
⬆️  Migrated environment to version 1 and upgraded all packages for environment 'name'.
```

## `pull`

`pull` only attempts to perform a migration if a user tries to pull an
environment onto a system not supported by that environment.

The user is prompted whether they want to both migrate the manifest and
add their system:
```
! The environment you are trying to pull needs to be migrated to the latest
environment format, and it's not yet compatible with your system (aarch64-darwin).

> Don't pull this environment.
  Pull this environment anyway, migrate it to version 1, and add 'aarch64-darwin' to the supported systems list.
[Use 'flox pull --force' to automatically add your system to the list of compatible systems]
```

If migration succeeds, a confirmation is printed:
```
⬆️  Migrated environment to version 1.
✨ Pulled owner/name from https://hub.flox.dev/

You can activate this environment with 'flox activate'
```

If migration fails, the user is directed to `flox edit`:
```
⚠️  could not automatically migrate manifest to version 1: unknown field `prefer-pre-releases`, expected `allow-pre-releases`
in `options.semver`

✨ Pulled owner/name from https://hub.flox.dev/

Migrated the environment to version 1 and included your system but could not
build.
Use 'flox edit' to address issues before activating.
```

---------

Co-authored-by: Dan Carley <dan.carley@gmail.com>