chore(show): Remove deprecated --all argument (#1660)
## Proposed Changes

Remove the warning that was added in d27e0d00 so that it will now
produce an error:

    % flox show --all
❌ ERROR: expected `<pkg-path>`, got `--all`. Pass `--help` for usage
information

This has been in two releases (1.0.6 and 1.0.7) and the timing is good
to remove it for 1.1.0 as a breaking change.

## Release Notes

Removed the deprecated `--all` flag from `flox show` now that it's the
default.