feat: add `options.cuda-support` to manifest
Instead of the secret feature flag FLOX_FEATURES_ENV_ENABLE_CUDA, add
`options.cuda-support` to the manifest to control the CUDA activation
script. Default to enabling CUDA if the flag is not set, which is change
in behavior since we currently default to false.

Add a `manifest()` function to the `Environment` trait to simplify
getting a typed manifest for an environment.