test(functional): Combine with integration tests
We agreed to run the functional tests on every commit, rather than just
the merge queue and on the `main` branch, because we've been frequently
surprised by the tests failing late in the development cycle.

I'm taking a slightly more radical approach by combining them back into
the integration tests because they aren't that fundamentally different
and it seems that they were only split originally because of the extra
runtime.

I've renamed the files to make it slightly clearer what they're doing.