feat: migrate hook.script to hook.on-activate (#1656)
When performing migration of manifest version 0 to version 1, rename
hook.script to hook.on-activate if hook.on-activate doesn't already
exist.

Also add formatter for MigrationError.

MigrationError wraps some errors that should be pretty printed, but the
wrapping causes the formatting function to not be printed. Fix this by
adding a formatting function that calls the formatters for the contained
errors.

Errors were being printed as:
```
❌ ERROR: Upgrade failed.: could not automatically migrate manifest to version 1: unknown field `prefer-pre-releases`, expected `allow-pre-releases`
in `options.semver`
```

But they should be printed as:
```
❌ ERROR: Could not automatically migrate manifest to version 1:

unknown field `prefer-pre-releases`, expected `allow-pre-releases`

Use 'flox edit' to resolve errors and then try again.
```

---------

Co-authored-by: Dan Carley <dan.carley@gmail.com>