test(integration): Use prod catalog not preview
The preview instance is currently broken:

    % FLOX_CATALOG_URL=https://flox-catalog.preview.flox.dev flox show hello
    ❌ ERROR: unexpected catalog connection error: Unexpected Response: Response { url: Url { scheme: "https", cannot_be_a_base: false, username: "", password: None, host: Some(Domain("flox-catalog.preview.flox.dev")), port: None, path: "/api/v1/catalog/packages/hello", query: Some("page=0&pageSize=10"), fragment: None }, status: 502, headers: {"server": "nginx", "date": "Mon, 24 Jun 2024 10:25:54 GMT", "content-type": "text/html", "content-length": "150", "connection": "keep-alive"} }

Besides unblocking changes to `flox/flox` right now it would be
beneficial to use production because that instance should be more
stable (e.g. we're less likely to make breaking changes) and is
representative of what a release needs to be compatible with after the
initial 1.1 release.

Also this value doesn't need to be a managed secret. In fact, it's
clearer if it's not, because you can reason about the value without
looking in other repos.