test(cross-system): Support pull on aarch64-darwin
To support changing the Mac runner from x64 to ARM64. This test was
silently skipped because we didn't have a matching case.

We can't use `get_system_other_than_current` because we don't push for
`aarch64-linux` and the test is intended to skip rather than fail when
the job runs on the opposing platform.