refactor: block imperative commands using errors
Previously, when a local environment was not in sync with the current generation
while executing an imperative command, flox would `panic`.
To integrate better with our error handling mechanisms,
this adds an `ManagedEnvironmentError::CheckoutOutOfSync` error variant
that is returned instead of panicking and can be matched against in downstream code.