test(cuda): Integration test CUDA detection
Add new integration tests for the existing CUDA detection functionality.

In the long term it would be preferable to observe the end-to-end
behaviour of using PyTorch on an actual GPU instance but we don't
currently have the CI setup to do so and we need to be careful about
limiting costs.

This runs with Matthew's idea of passing in an FHS root that we can
treat as a "sandbox" to stub system paths inside. I considered using
`chroot` (or similar) to do the same but it requires root, some hoop
jumping to work on MacOS, and I suspect even more hoop jumping to work
with Flox/Nix.