fix(cuda): Fix detection in other dirs
This fixes the last test which was failing. Previously the script, when
run with `set -e` from the test, would exit on find's non-zero exit code
if any of the preceding directories didn't exist.

So if `/run/opengl-drivers` didn't exist then we wouldn't check any of
the other system dirs. I couldn't see a `find` argument to do this so
I've had to catch any non-zero exit from `find`.