tests: Combine functional tests with integration tests (#1687)
## Proposed Changes

This is best reviewed commit-by-commit.

---

**chore(justfile): Remove integ-file**

This is the same as `integ-tests` and doesn't work as advertised, so
just reference the ability to run a single test file by the relative
path to the tests directory.

**docs(CONTRIBUTING): Update integration test docs**

Simplify them by:

- assuming that we're in `nix develop` by the time we get this far down
  the doc, like the other sections do
- run everything through `just` since it handles the arguments and
  building for us in a consistent way
- standardise on the integration (rather than functional) tests since
  they exercise more functionality and we're going to combine them in a
  following commit

It's worth noting that the `--watch` argument no longer appears to work,
seemingly because `flox-cli-tests` makes a copy of the test files before
running, but I'm not going to look further into this right now

**docs(cli/tests): Link to CONTRIBUTING**

These docs are out-of-date and duplicate content that is in the
contributing guide, so link to there instead.

**chore(ci): Fix name of build bats job**

This uses `flox-cli-tests` rather than `flox-tests`.

**test(functional): De-dupe support files**

Use the centralised and namespaced copies of these files, which are
otherwise identical.

**test(functional): Merge support files back**

Copy any differences from `cli/tests/` to verify that the functional
tests will work and highlight that we don't lose any functionality when
the `tests/` directory is deleted. The purpose of this commit is just to
make code review easier.

**test(catalog): Remove duplicate upgrade test**

We already cover this test case in `upgrade.bats`. This test case was
the odd-one-out in the functional tests because it uses mocks when the
other use the live catalog server.

**test(functional): Combine with integration tests**

We agreed to run the functional tests on every commit, rather than just
the merge queue and on the `main` branch, because we've been frequently
surprised by the tests failing late in the development cycle.

I'm taking a slightly more radical approach by combining them back into
the integration tests because they aren't that fundamentally different
and it seems that they were only split originally because of the extra
runtime.

I've renamed the files to make it slightly clearer what they're doing.

## Release Notes

N/A, dev only facing.