chore(cuda): Add to manifest template
Add a commented entry to the manifest template so that users can easily
find the option if they run into any problems with CUDA detection.

It's a little bit odd to add a key that then has a prefix to comment it
out. It's possible that `toml_edit` could change this behaviour in the
future but:

- we have unit tests around it
- we're serialising it straight away and the deserialised version of
  that will pass it as a comment

The only other options I could find were:

- add the suffix to the `options` table which puts it above `systems`
  and I think that should still be more prominent
- add the suffix to the `systems` key which feels even odder