fix: skip_serializing_none for manifest and lock (#1708)
Add `skip_serializing_none` to types for the manifest and lock.

Currently, if a newer version of flox has an optional field, if an
environment is locked, the field is put as null in the lockfile. If an
older version of flox without that field attempts to edit the
environment, it will throw an error, e.g:
```
❌ ERROR: unrecognized manifest field 'options.cuda-detection'
```

Since serializing null doesn't gain us anything, we might as well skip
it. It will also make lockfiles slightly smaller.