feat: build flox-activate package distinct from flox-pkgdb (#1709)
## Proposed Changes

The activation package was previously assembled at build time from a set
of static files (etc-profiles) and other scripts created by pkgdb, then
it evolved into a set of static files with package path substitutions
performed at Nix build time. This introduced the wrinkle that changes to
the flox-activate package could not be picked up without having to exit
and re-enter the `nix develop` shell.

This change splits out the flox-activate package as distinct from
flox-pkgdb so that the package can be quickly re-rendered with an
invocation of `nix build`. It then updates the Makefile, flox-pkgdb and
flox-cli builds to use the files from this package.

## Release Notes

N/A

---------

Co-authored-by: Michael Brantley <brantley@flox.com>