test(catalog-live): Move env and always run
Move the environment variable into the tests, rather than sourcing it
from GitHub, so that we always run the tests. This makes the behaviour
of the integration tests consistent across:

- local dev
- `cli-dev` job in CI
- `nix-build-bats-tests` in CI

Previously the latter was the only job to run these which makes it hard
to reason and work on failures.

You could still focus or skip these temporarily if you're working
somewhere without an internet connection.