test(catalog-live): Fix and always run (#1735)
## Proposed Changes

**test(catalog-live): Move env and always run**

Move the environment variable into the tests, rather than sourcing it
from GitHub, so that we always run the tests. This makes the behaviour
of the integration tests consistent across:

- local dev
- `cli-dev` job in CI
- `nix-build-bats-tests` in CI

Previously the latter was the only job to run these which makes it hard
to reason and work on failures.

You could still focus or skip these temporarily if you're working
somewhere without an internet connection.

**test(catalog-live): Relax package description**

The tests have failed in CI because the package description in the
latest catalog scrape has changed from:

    A program that produces a familiar, friendly greeting

To:

    Program that produces a familiar, friendly greeting

Relax the assertion so that we still find the right package but are more
lenient about the contents.

## Release Notes

N/A