fix: abort when failing to build flox-activation-scripts package
The previous modification correctly rebuilt the flox-activation-scripts
package when necessary, but didn't detect when that build failed, so
processing continued until some other test failed as the result of the
ACTIVATION_SCRIPTS_DIR variable being defined but empty.

This patch changes the logic so that we immediately abort the pkgdb
build in the event that we are unable to build the flox-activation-scripts
package.