test(services): Don't bg+sleep+cleanup
We don't need to background `process-compose`, wait for it to have
started, and then cleanup afterwards now that we've switched to using a
non-network "service" because `process-compose` will exit on completion.

It's wrapped in a 2s timeout, which will issue a kill signal if
exceeded, to prevent the test hanging if `process-compose` does
something unexpected.

We still set the socket, even though we don't access it, because it
exercises the CLI code path for setting the environment variable and if
not set `process-compose` will default to binding a TCP socket which
will conflict with multiple runs on the same host.