chore: remove `ManifestInstall::catalog_descriptors_ref`
Avoid splitting descriptor methods between `ManifestInstall` and `TypedManifestCatalog` impls.
Currently there are a number of higher level methods implemented for `TypedManifestCatalog`,
that return subsets of descriptors from the contained `ManifestInstall` instance.

For consistency removing the methods from `ManifestInstall`
and moving it to `TypedManifestCatalog` were considered.
As currently it's unclear whether we will need specific aggregators
for the individual variants of `ManifestPackageDescriptor`,
and filtering using `ManifestPackageDescriptor::as_*_descriptor_ref` is rather simple
I opted to omit this method for now.

Tangentially, I also believe that these methods currently on `TypedManifestCatalog`
should rather be implemented on `ManifestInstall`,
but that's obviously too large of a change for this change set for this change set
and more of a personal preference than a necessity.