chore: Increase catalog page size from 10 to 1000 (#1730)
We now retrieve up to 1000 results at a time while paging, increased
from
only 10. The only cases where we actually use pagination (search --all
and show) retrieve the entire set of results anyway, so smaller pages
don't really reduce the work and substantially increase the latency. For
ordinary search, where we limit the number of results we show, we still
respect that limit and only retrieve the necessary number of rows.