feat: generate service config file at build time (#1712)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
This generates the service config file at build time and stores it in
the environment directory alongside the `activate` script.

Other notable changes:
- Replaced the `serde_yaml` dependency with `serde_yml`. `serde_yaml` is
no longer maintained and `serde_yml` is a fork that's maintained.
- Added a `features` field to `Flox` since feature detection is
necessary in `flox-rust-sdk` and `Flox` seems like the right place to
store config.
- Sets `_FLOX_SERVICES_SOCKET` in the activation rather than allowing
passing from the outside.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A

<!-- Many thanks! -->

---------

Co-authored-by: Dan Carley <dan.carley@gmail.com>
Co-authored-by: Matthew Kenigsberg <matthew@floxdev.com>