chore(envs): Remove implicit build from link()
To make a clearer separation between them, which will also allow us to
remove some conditional and duplicated logic. For example we no longer
need to build the service config in `link` because we know that we're
only ever dealing with an already built environment.

Previously:

- build would: always build and never link
- link would: maybe build and always link

Now:

- build will: always build and never link
- link will: never build and always link

As far as I can tell, the only times that `store_path` would be `None`
is when we've attempted a transaction on a temporary environment and
found that we don't need the resulting build, such as when a package is
already installed.

So they have been wrapped with conditionals further up the call chain.
Arguably this has resulted in a few more lines of code but is easier to
reason about closer to the original call site.