refactor(env): Move pkgdb build/link up to CoreEnv
Up from `LockedManifest` to `CoreEnvironment`. This has two benefits:

1. Simplifies the argument construction now that `link` no longer
implicitly builds because we know that `out-link` and `store-path` have
to be passed together.

2. Allows us to pass the original lockfile path rather than having to
read and parse it, which is both simpler and also fixes the bug in #1706
when the lock contents are too large to pass as a CLI string argument.

I've moved up all of the related errors; believe I caught them all.

I did attempt to nest the conditionals in `build` for constructing
`service_config_path` but it resulted in the tempdir being dropped
before the `pkgdb` command was called.