refactor(env): Move pkgdb logging down
Always debug log the `pkgdb` command before we call it. In doing so, we
can remove a lot of surplus logging further up the call chain.

The `pkgdb` command contains all of the information we need, in a much
more concise format, now that the preceding commits have:

- separated the `buildenv` and `linkenv` commands
- pass a lockfile path rather than the string contents

I've left the function signatures accepting a `&Flox` for now for
symmetry with `build` and so that we don't have to churn code on all of
the callers.