refactor(services): Simplify config tempfile
The temporary service config file must not be cleaned up until it's used
by `pkgdb buildenv`. Previously this was ensured by carrying the parent
`TempDir` in a struct with it. Simplify this by instead using
`NamedTempFile`, calling `keep`, and letting it live until the parent
`flox.temp_dir` is cleaned up for us.