chore: refactor ld-floxlib into separate package (#1742)
## Proposed Changes

Refactor the build of `ld-floxlib.so` into a separate package, and
establish it strictly as a requisite of the `flox-activation-scripts`
repository (which is the one place from which it is referenced).

Also take this opportunity to create `test` Makefile target and unit
tests to be invoked in the Nix sandbox during the build.

## Release Notes

N/A

---------

Co-authored-by: Michael Brantley <brantley@flox.com>