fix: print environment description in `delete` (#1748)
55ff6afd2c87eb1c729ac06cfb67a995957b34ae shortened the prompt for `flox
delete` from:
```
You are about to delete your environment {description}. Are you sure?
```
to
```
Are you sure?
```

This makes sense for cases where the environment is explicitly specified
with `-d`, but because we have active environment detection, `flox
delete` can delete an environment not in the current directory, and the
prompt doesn't give you any signal that's happening:
```
> flox delete
! Are you sure? (y/N)
[Use `-f` to force deletion]
```

When `-d` is not passed, print out a message with the full description.