refactor(services): Simplify config tempfile (#1750)
## Proposed Changes

Code review suggestion from:
https://github.com/flox/flox/pull/1712#discussion_r1666995321

The temporary service config file must not be cleaned up until it's used
by `pkgdb buildenv`. Previously this was ensured by carrying the parent
`TempDir` in a struct with it. Simplify this by instead using
`NamedTempFile`, calling `keep`, and letting it live until the parent
`flox.temp_dir` is cleaned up for us.

## Release Notes

N/A

---------

Co-authored-by: Zach Mitchell <zmitchell@fastmail.com>