feat(services): Scaffolding for stop command
This is the first of the imperative service commands that are nested
under `flox services <cmd> [args]`.

It is currently hidden from the `--help` output until we enable it by
default. I couldn't see an easy way to tie this behaviour to the feature
flag when using `bpaf_derive` but it also doesn't seem like a big deal.

TODOs about telemetry naming need to be resolved before launch. The only
other nested commands I can see at the moment is `flox auth` which
sidesteps the question by grouping everything under `auth` but we'll
want more granular telemetry for services.