test(services): Scaffolding for stop tests
I'm currently calling `process-compose` directly to perform the
imperative actions and then we can swap in `flox services stop` commands
as they get implemented.

We need to do some gymnastics to start, stop, and inspect
`process-compose` until we have implemented other services
functionality. These can also be swapped out as we go.

I've structured the tests with `activate -- bash <(cat <<'EOF'` and
sourced the start/cleanup script so that we can re-use it within each
activation and keep the commands that are unique to each test case more
prominent than they would be if each one called an individual script.

Needing to do everything within a single activation, so that we cleanup
`process-compose` afterwards, is going to make inspecting the output of
`flox services` and `process-compose process` a bit awkward for now.