feat(services): Prevent outside activation
Prevent using any `flox services …` commands if you're not within an
activation that has services where the socket environment variable has
been set, which is currently controlled by the feature flag and having a
`services` section in the manifest.

There is no guarantee that services have actually been started and that
the socket has been created, which we'll need to handle errors for in
the provider regardless.

This doesn't prevent you calling `--help`.

I've implemented this in the outer `flox services` delegator so that we
don't have to implement it within each sub-command.