refactor: Remove _FLOX_SERVICES_CONFIG_PATH
I believe the original intention of this was that we would be able to
test pass through a `process-compose` config from integration tests
without going through an entire environment and manifest but I haven't
found this useful when writing the `stop` tests because the socket
environment variable in an activation is gated on there being services
in the manifest. So we might as well just test the whole callpath.