feat(services): Inject flox_never_exit process
To prevent `process-compose` from exiting when all services have been
stopped. This workaround makes it easier to test right now, but will
also make to handle the complete lifecycle from `activate` to exit
because we won't need to keep checking whether `process-compose` is
running and start it again.

I've implemented the injection on the serialization and stripped it off
in deserialization so that it doesn't leak into other representations.
This will make it easier to iterate over all services from the manifest
but we'll need to account for it if we iterate over `process-compose
list` output.

I had to move the existing `touch_file` test over to the same format as
the `stop` tests because we can no longer call `process-compose` as a
one-shot command on the generated config.