feat: start services on activation with flag (#1760)
## Proposed Changes

<!-- Describe the changes proposed in this pull request. -->
<!-- Please provide links to any issue(s) which are expected to be
resolved. -->
With the services feature flag you can now automatically start services
during activation via `flox activate --start-services` or `flox activate
-s`. The services are not shut down automatically, so you'll need to
call `process-compose down` with the appropriate arguments in order to
clean up services.

Another notable change is that the check for an unsupported shell is now
done closer to the start of the `activate` script. The existing checks
are left in as an extra layer of protection in case there's a bug in the
Bash logic.

## Release Notes

<!-- Describe any user facing changes. Use "N/A" if not applicable. -->
N/A


<!-- Many thanks! -->