feat(services): Implement stop all
Calling out to `process-compose list` for the current services, minus
our "never exit" service, because it proved to be easier than getting a
definitive list from the environment.

We could later use ProcessStates to check whether a service is running
before attempting to stop it.

Required borrowing the socket path now that we use it in multiple places.

The output format came from:

- https://github.com/F1bonacc1/process-compose/blob/v1.6.1/src/types/process.go#L89-L100