feat: add helper-functions.bash collection w/ flox_portgrab()
This patch adds a new place for bash helper functions to be used within
the on-activate hook, and uses it to add a new flox_portgrab() function.
This function is useful in the service management area for grabbing an
ephemeral port that is guaranteed to be free and available for use.

Example usage:

SERVER_PORT=$(flox_portgrab)